chore(placement): move out of toc
Also add it to the search result page.
