+++ client/src/document/organisms/toc/index.tsx
@@ -3,7 +3,7 @@ import React, { useState } from "react";
 import "./index.scss";
 import { Toc } from "../../../../../libs/types/document";
 import { useFirstVisibleElement } from "../../hooks";
-import { Placement } from "./placement";
+import { Placement } from "../../../ui/organisms/placement";
 import { PLACEMENT_ENABLED } from "../../../env";
 
 export function TOC({ toc }: { toc: Toc[] }) {
